Modify divide() to return a float

+++ Math.cpp
@@ -17,7 +17,7 @@ int multiply(int x, int y)
 	return x * y;
 }
 
-int divide(int x, int y)
+float divide(int x, int y)
 {
-	return x / y;
+	return float(x) / float(y);
 }
