+++ math.h
@@ -3,4 +3,4 @@
 int add(int x, int y);
 int subtract(int x, int y);
 int multiply(int x, int y);
-int divide(int x, int y);
+float divide(int x, int y);
